state_mgmt: split SessionInfo and dummy-final-classification out of session_state.py
Step 5 of the simplification plan - pure file organization, no behaviour change.
Both moved verbatim.

SessionInfo -> session_info.py, plus a _buildPitTimeLossDict helper that dedups the
previously-duplicated F1/F2 pit-loss-dict construction. No other file imports
SessionInfo directly - the 9 external consumers all go through a SessionState
instance attribute.

Dummy-final-classification builders -> dummy_final_classification.py as a
DummyFinalClassificationMixin, combined into SessionState via multiple inheritance
so the method bodies keep their exact self.-based form. TYPE_CHECKING-guarded
attribute declarations let a type checker resolve self.m_num_active_cars /
self.m_session_info inside the mixin's own methods without a circular import or
runtime cost - same pattern to reuse for the remaining two mixins (packet
processors, analysis helpers) still to come.
